requirements.txt
# Testing
django-nose==1.2
nose==1.3.4
spec==0.11.1
coverage>=3.7,<3.8
coveralls==0.2
tox==1.7.0
flake8==2.2.5

# Docs
Sphinx==1.2.3